+++ compiler-plugin/src/main/java/io/ballerina/stdlib/http/compiler/HttpResourceValidator.java
@@ -438,9 +438,30 @@ class HttpResourceValidator {
             return isValidPayloadParamType(arrTypeSymbol, ctx, paramLocation, paramName);
         } else if (kind == TypeDescKind.TYPE_REFERENCE) {
             typeDescriptor = ((TypeReferenceTypeSymbol) typeDescriptor).typeDescriptor();
-            TypeDescKind typeDescKind = retrieveEffectiveTypeDesc(typeDescriptor);
-            return typeDescKind == TypeDescKind.RECORD || isValidPayloadParamType(typeDescriptor, ctx, paramLocation,
-                                                                                  paramName);
+            String typeName = typeDescriptor.signature();
+            TypeDescKind typeDescKind = typeDescriptor.typeKind();
+            if (typeDescKind == TypeDescKind.INTERSECTION) {
+                typeDescriptor = getEffectiveTypeFromReadonlyIntersection((IntersectionTypeSymbol) typeDescriptor);
+                if (typeDescriptor == null) {
+                    reportInvalidIntersectionType(ctx, paramLocation, typeName);
+                    return true;
+                }
+                typeDescKind = typeDescriptor.typeKind();
+            }
+            if (typeDescKind == TypeDescKind.RECORD) {
+                Map<String, RecordFieldSymbol> recordFieldSymbols =
+                        ((RecordTypeSymbol) typeDescriptor).fieldDescriptors();
+                recordFieldSymbols.forEach((key, value) -> {
+                    TypeSymbol fieldTypeSymbol = value.typeDescriptor();
+                    if (!isValidPayloadParamType(fieldTypeSymbol, ctx, paramLocation, paramName)) {
+                        reportInvalidRecordFieldType(ctx, paramLocation, paramName, fieldTypeSymbol.signature());
+                    }
+                });
+                return true;
+            } else {
+//                typeDescriptor = ((TypeReferenceTypeSymbol) typeDescriptor).typeDescriptor();
+                return isValidPayloadParamType(typeDescriptor, ctx, paramLocation, paramName);
+            }
         } else if (kind == TypeDescKind.MAP) {
             typeDescriptor = ((MapTypeSymbol) typeDescriptor).typeParam();
             return isValidPayloadParamType(typeDescriptor, ctx, paramLocation, paramName);
@@ -878,5 +899,10 @@ class HttpResourceValidator {
 
     private static void reportInvalidHeaderRecordRestFieldType(SyntaxNodeAnalysisContext ctx, Location location) {
         updateDiagnostic(ctx, location, HttpDiagnosticCodes.HTTP_144);
+    }
+
+    private static void reportInvalidRecordFieldType(SyntaxNodeAnalysisContext ctx, Location location,
+                                                     String paramName, String typeName) {
+        updateDiagnostic(ctx, location, HttpDiagnosticCodes.HTTP_145, typeName, paramName);
     }
 }